# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/groups": {
      "post": {
        "operationId": "GroupController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "groups"
        ]
      },
      "get": {
        "operationId": "GroupController_list",
        "parameters": [
          {
            "name": "public",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mine",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedGroupsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "groups"
        ]
      }
    },
    "/groups/{id}": {
      "get": {
        "operationId": "GroupController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "groups"
        ]
      }
    },
    "/groups/{id}/join": {
      "post": {
        "operationId": "GroupController_join",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupMemberResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "groups"
        ]
      }
    },
    "/groups/{id}/members/me": {
      "delete": {
        "operationId": "GroupController_leave",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "groups"
        ]
      }
    },
    "/groups/{id}/members": {
      "get": {
        "operationId": "GroupController_listMembers",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedGroupMembersResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "groups"
        ]
      }
    },
    "/challenges": {
      "post": {
        "operationId": "ChallengeController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChallengeResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "challenges"
        ]
      },
      "get": {
        "operationId": "ChallengeController_list",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mine",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedChallengesResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "challenges"
        ]
      }
    },
    "/challenges/{id}": {
      "get": {
        "operationId": "ChallengeController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChallengeResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "challenges"
        ]
      }
    },
    "/challenges/{id}/join": {
      "post": {
        "operationId": "ChallengeController_join",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChallengeMemberResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "challenges"
        ]
      }
    },
    "/challenges/{id}/members/me": {
      "delete": {
        "operationId": "ChallengeController_leave",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "challenges"
        ]
      }
    },
    "/challenges/{id}/members": {
      "get": {
        "operationId": "ChallengeController_listMembers",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedChallengeMembersResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "challenges"
        ]
      }
    },
    "/challenges/{id}/progress": {
      "get": {
        "operationId": "ChallengeController_getMyProgress",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChallengeProgressResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "challenges"
        ]
      }
    },
    "/challenges/{id}/complete": {
      "post": {
        "operationId": "ChallengeController_complete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "challenges"
        ]
      }
    },
...
  },
  "info": {
    "title": "Habit Tracker API",
    "description": "API for habit tracking, community, challenges and gamification",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "GroupResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "object"
          },
          "coverImageUrl": {
            "type": "object"
          },
          "profileImageUrl": {
            "type": "object"
          },
          "isPublic": {
            "type": "boolean"
          },
          "allowMemberInvites": {
            "type": "boolean"
          },
          "memberCount": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "isPublic",
          "allowMemberInvites",
          "createdAt",
          "updatedAt"
        ]
      },
      "PaginatedGroupsResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GroupResponseDto"
            }
          },
          "nextCursor": {
            "type": "string"
          }
        },
        "required": [
          "items"
        ]
      },
      "GroupMemberResponseDto": {
        "type": "object",
        "properties": {
          "groupId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "joinedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "groupId",
          "userId",
          "role",
          "joinedAt"
        ]
      },
      "PaginatedGroupMembersResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GroupMemberResponseDto"
            }
          },
          "nextCursor": {
            "type": "string"
          }
        },
        "required": [
          "items"
        ]
      },
      "ChallengeResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "groupId": {
            "type": "string"
          },
          "creatorId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "object"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "COMPLETED",
              "CANCELLED"
            ]
          },
          "imageUrl": {
            "type": "object"
          },
          "taskTemplate": {
            "type": "object"
          },
          "startDate": {
            "format": "date-time",
            "type": "string"
          },
          "endDate": {
            "type": "object"
          },
          "autoCreateTaskOnJoin": {
            "type": "boolean"
          },
          "onTrackStreakThreshold": {
            "type": "number"
          },
          "memberCount": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "groupId",
          "creatorId",
          "title",
          "status",
          "taskTemplate",
          "startDate",
          "autoCreateTaskOnJoin",
          "onTrackStreakThreshold"
        ]
      },
      "PaginatedChallengesResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChallengeResponseDto"
            }
          },
          "nextCursor": {
            "type": "string"
          }
        },
        "required": [
          "items"
        ]
      },
      "ChallengeMemberResponseDto": {
        "type": "object",
        "properties": {
          "challengeId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "currentStreak": {
            "type": "number"
          },
          "joinedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "challengeId",
          "userId",
          "currentStreak",
          "joinedAt"
        ]
      },
      "PaginatedChallengeMembersResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChallengeMemberResponseDto"
            }
          },
          "nextCursor": {
            "type": "string"
          }
        },
        "required": [
          "items"
        ]
      },
      "ChallengeProgressResponseDto": {
        "type": "object",
        "properties": {
          "currentStreak": {
            "type": "number"
          },
          "onTrack": {
            "type": "boolean"
          },
          "requiredStreak": {
            "type": "number"
          }
        },
        "required": [
          "currentStreak",
          "onTrack",
          "requiredStreak"
        ]
      },
...
    }
  }
}