# src/modules/community/application/dtos/community-response.dto.ts
import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';
import { PostType, PostVisibility, ChallengeStatus } from '../../domain/entities/community.entity';

export class PostResponseDto {
  @ApiProperty()
  id!: string;

  @ApiProperty()
  userId!: string;

  @ApiPropertyOptional()
  groupId!: string | null;

  @ApiPropertyOptional()
  challengeId!: string | null;

  @ApiPropertyOptional()
  content!: string | null;

  @ApiProperty({ enum: PostType })
  type!: PostType;

  @ApiPropertyOptional()
  mediaUrl!: string | null;

  @ApiProperty({ enum: PostVisibility })
  visibility!: PostVisibility;

  @ApiProperty()
  isAutoGenerated!: boolean;

  @ApiPropertyOptional()
  challengeTitle!: string | null;

  @ApiProperty({ type: [String] })
  taggedUserIds!: string[];

  @ApiProperty({ type: [String] })
  hashtags!: string[];

  @ApiProperty()
  commentCount!: number;

  @ApiPropertyOptional()
  likeCount?: number;

  @ApiProperty()
  createdAt!: Date;

  @ApiProperty()
  updatedAt!: Date;
}

export class CommentResponseDto {
  @ApiProperty()
  id!: string;

  @ApiProperty()
  postId!: string;

  @ApiProperty()
  userId!: string;

  @ApiProperty()
  content!: string;

  @ApiProperty()
  createdAt!: Date;
}

export class ChallengeResponseDto {
  @ApiProperty()
  id!: string;

  @ApiProperty()
  groupId!: string;

  @ApiProperty()
  creatorId!: string;

  @ApiProperty()
  title!: string;

  @ApiPropertyOptional()
  description!: string | null;

  @ApiProperty({ enum: ChallengeStatus })
  status!: ChallengeStatus;

  @ApiPropertyOptional()
  imageUrl!: string | null;

  @ApiProperty()
  taskTemplate!: unknown;

  @ApiProperty()
  startDate!: Date;

  @ApiPropertyOptional()
  endDate!: Date | null;

  @ApiProperty()
  autoCreateTaskOnJoin!: boolean;

  @ApiProperty()
  onTrackStreakThreshold!: number;

  @ApiPropertyOptional()
  memberCount?: number;
}

export class ChallengeMemberResponseDto {
  @ApiProperty()
  challengeId!: string;

  @ApiProperty()
  userId!: string;

  @ApiProperty()
  currentStreak!: number;

  @ApiProperty()
  joinedAt!: Date;
}

export class GroupResponseDto {
  @ApiProperty()
  id!: string;

  @ApiProperty()
  name!: string;

  @ApiPropertyOptional()
  description!: string | null;

  @ApiPropertyOptional()
  coverImageUrl!: string | null;

  @ApiPropertyOptional()
  profileImageUrl!: string | null;

  @ApiProperty()
  isPublic!: boolean;

  @ApiProperty()
  allowMemberInvites!: boolean;

  @ApiPropertyOptional()
  memberCount?: number;

  @ApiProperty()
  createdAt!: Date;

  @ApiProperty()
  updatedAt!: Date;
}

export class GroupMemberResponseDto {
  @ApiProperty()
  groupId!: string;

  @ApiProperty()
  userId!: string;

  @ApiProperty()
  role!: string;

  @ApiProperty()
  joinedAt!: Date;
}

export class PaginatedPostsResponseDto {
  @ApiProperty({ type: [PostResponseDto] })
  items!: PostResponseDto[];

  @ApiPropertyOptional()
  nextCursor?: string;
}

export class PaginatedCommentsResponseDto {
  @ApiProperty({ type: [CommentResponseDto] })
  items!: CommentResponseDto[];

  @ApiPropertyOptional()
  nextCursor?: string;
}

export class PaginatedChallengesResponseDto {
  @ApiProperty({ type: [ChallengeResponseDto] })
  items!: ChallengeResponseDto[];

  @ApiPropertyOptional()
  nextCursor?: string;
}

export class PaginatedChallengeMembersResponseDto {
  @ApiProperty({ type: [ChallengeMemberResponseDto] })
  items!: ChallengeMemberResponseDto[];

  @ApiPropertyOptional()
  nextCursor?: string;
}

export class PaginatedGroupsResponseDto {
  @ApiProperty({ type: [GroupResponseDto] })
  items!: GroupResponseDto[];

  @ApiPropertyOptional()
  nextCursor?: string;
}

export class PaginatedGroupMembersResponseDto {
  @ApiProperty({ type: [GroupMemberResponseDto] })
  items!: GroupMemberResponseDto[];

  @ApiPropertyOptional()
  nextCursor?: string;
}

export class ChallengeProgressResponseDto {
  @ApiProperty()
  currentStreak!: number;

  @ApiProperty()
  onTrack!: boolean;

  @ApiProperty()
  requiredStreak!: number;
}