# src/modules/community/presentation/challenge.controller.ts
import { ApiTags, ApiOkResponse, ApiCreatedResponse, ApiNoContentResponse } from '@nestjs/swagger';
import { Body, Controller, Delete, Get, Param, Post, Query, UseGuards } from '@nestjs/common';
import { SessionGuard } from '../../../shared/infrastructure/auth/guards/session.guard';
import { CurrentUser } from '../../../shared/infrastructure/auth/decorators/current-user.decorator';
import type { AuthenticatedUser } from '../../../shared/domain/auth.types';
import { CreateChallengeUseCase } from '../application/create-challenge.use-case';
import { CreateChallengeDto } from '../application/dtos/create-challenge.dto';
import { ListChallengesUseCase } from '../application/list-challenges.use-case';
import { GetChallengeUseCase } from '../application/get-challenge.use-case';
import { JoinChallengeUseCase } from '../application/join-challenge.use-case';
import { LeaveChallengeUseCase } from '../application/leave-challenge.use-case';
import { ListChallengeMembersUseCase } from '../application/list-challenge-members.use-case';
import { GetChallengeProgressUseCase } from '../application/get-challenge-progress.use-case';
import { CompleteChallengeUseCase } from '../application/complete-challenge.use-case';
import {
  ChallengeResponseDto,
  ChallengeMemberResponseDto,
  PaginatedChallengesResponseDto,
  PaginatedChallengeMembersResponseDto,
  ChallengeProgressResponseDto,
} from '../application/dtos/community-response.dto';

@ApiTags('challenges')
@Controller('challenges')
export class ChallengeController {
  constructor(
    private readonly createChallengeUseCase: CreateChallengeUseCase,
    private readonly listChallengesUseCase: ListChallengesUseCase,
    private readonly getChallengeUseCase: GetChallengeUseCase,
    private readonly joinChallengeUseCase: JoinChallengeUseCase,
    private readonly leaveChallengeUseCase: LeaveChallengeUseCase,
    private readonly listChallengeMembersUseCase: ListChallengeMembersUseCase,
    private readonly getChallengeProgressUseCase: GetChallengeProgressUseCase,
    private readonly completeChallengeUseCase: CompleteChallengeUseCase,
  ) {}

  @Post()
  @UseGuards(SessionGuard)
  @ApiCreatedResponse({ type: ChallengeResponseDto })
  async create(
    @CurrentUser() user: AuthenticatedUser,
    @Body() body: Omit<CreateChallengeDto, 'userId'>,
  ) {
    return this.createChallengeUseCase.execute({
      ...body,
      userId: user.id,
    });
  }

  @Get()
  @ApiOkResponse({ type: PaginatedChallengesResponseDto })
  async list(
    @Query('groupId') groupId?: string,
    @Query('userId') userId?: string,
    @Query('mine') mineStr?: string,
    @Query('limit') limitStr?: string,
    @Query('cursor') cursor?: string,
    @CurrentUser() user?: AuthenticatedUser,
  ) {
    const limit = Math.min(Math.max(parseInt(limitStr ?? '20', 10) || 20, 1), 100);
    const mine = mineStr === 'true' || mineStr === '1';
    return this.listChallengesUseCase.execute({
      groupId,
      userId: mine && user?.id ? user.id : userId,
      limit,
      cursor,
    });
  }

  @Get(':id')
  @ApiOkResponse({ type: ChallengeResponseDto })
  async getById(@Param('id') id: string) {
    return this.getChallengeUseCase.execute({ challengeId: id });
  }

  @Post(':id/join')
  @UseGuards(SessionGuard)
  @ApiOkResponse({ type: ChallengeMemberResponseDto })
  async join(@Param('id') id: string, @CurrentUser() user: AuthenticatedUser) {
    return this.joinChallengeUseCase.execute({ challengeId: id, userId: user.id });
  }

  @Delete(':id/members/me')
  @UseGuards(SessionGuard)
  @ApiNoContentResponse()
  async leave(@Param('id') id: string, @CurrentUser() user: AuthenticatedUser) {
    await this.leaveChallengeUseCase.execute({ challengeId: id, userId: user.id });
  }

  @Get(':id/members')
  @ApiOkResponse({ type: PaginatedChallengeMembersResponseDto })
  async listMembers(
    @Param('id') id: string,
    @Query('limit') limitStr?: string,
    @Query('cursor') cursor?: string,
  ) {
    const limit = Math.min(Math.max(parseInt(limitStr ?? '20', 10) || 20, 1), 100);
    return this.listChallengeMembersUseCase.execute({ challengeId: id, limit, cursor });
  }

  @Get(':id/progress')
  @UseGuards(SessionGuard)
  @ApiOkResponse({ type: ChallengeProgressResponseDto })
  async getMyProgress(@Param('id') id: string, @CurrentUser() user: AuthenticatedUser) {
    return this.getChallengeProgressUseCase.execute({ challengeId: id, userId: user.id });
  }

  @Post(':id/complete')
  @UseGuards(SessionGuard)
  @ApiNoContentResponse()
  async complete(@Param('id') id: string, @CurrentUser() user: AuthenticatedUser) {
    await this.completeChallengeUseCase.execute({ challengeId: id, userId: user.id });
  }
}